Module tis_100_superoptimizer::TIS_100

The Tessellated Intelligence System is a

massively parallel computer architecture comprised of non-uniformly interconnected heterogeneous nodes.

The manual can be found online. Another useful resource is the unofficial TIS-100 Hackers Guide.

Examples

Here we take a node and execute some instructions and check if double 1 has the expected result.

extern crate tis_100_superoptimizer;

use tis_100_superoptimizer::TIS_100::{Node,Instruction,Source,Destination,Register};

fn main() {
    let node: Node = Node::new();
    let last: Node = node
        .execute(Instruction::MOV(Source::Literal(1), Destination::Register(Register::ACC))).unwrap()
        .execute(Instruction::ADD(Source::Register(Register::ACC))).unwrap();

    assert_eq!(2, last.acc);
}Run

Modules

Ports
Ports are the IO channels for a Node

Structs

Node
A Node models the basic execution node in TIS-100. You change a node state by running Programs on it or executing an Instruction on it.

Program
A Program is a sequence of Instructions

Enums

Cycle
Determines how many cycles to run a program

Destination
Destination are either ports or registers

ErrorStatus
The ErrorStatus a Program of an unsuccessful run on a certain Node

Instruction
Instructions are executed by a Node

Register
Different types of Registers known in TIS-100

Source
Source are either ports, registers or literals