Struct tis_100_superoptimizer::TIS_100::Node

pub struct Node {
    pub acc: i32,
    // some fields omitted
}

A Node models the basic execution node in TIS-100. You change a node state by executing an Instruction on it.

Fields
acc: i32

The accumulator for the basic execution node.

Methods

impl Node

fn new() -> Node

Create a Node with defaults for accumulator and backup registers

fn increment_pc(&self) -> Node

Create a Node from self with the program counter incremented

fn set_pc(&self, pc: isize) -> Node

Create a Node from self with a prescribed program counter value

fn set_acc(&self, acc: i32) -> Node

Create a Node from self with a prescribed accumulator register value

fn set_bac(&self, bac: i32) -> Node

Create a Node from self with a prescribed backup register value

fn execute(&self, instruction: Instruction) -> Node

Execute the instruction on this Node. Returns a Node that reflects the changes the instruction would have on this Node.