Struct tis_100_superoptimizer::TIS_100::Node

pub struct Node {
    pub acc: i32,
    pub up: Port,
    pub down: Port,
    // some fields omitted
}

A Node models the basic execution node in TIS-100. You change a node state by running Programs on it or executing an Instruction on it.

Fields
acc: i32

The accumulator for the basic execution node.

up: Port

The up Port used for reading

down: Port

The down Port used for writing

Methods

impl Node

fn new() -> Node

Create a Node with defaults for accumulator, backup registers, program counter and program

fn load(&self, program: Program) -> Node

Loads a program in this Node

fn run(&self, allowed_cycles: Cycle) -> Result<Node, ErrorStatus>

Run the loaded program, returning an calculation state

fn set_up(&self, up: Port) -> Node

Create a Node from self with a prescribed down port

fn execute(&self, instruction: Instruction) -> Option<Node>

Execute the instruction on this Node. Returns a Node that reflects the changes the instruction would have on this Node.